[#5] Custom argument parser
Problem: ContractRegistry is a bit too restrictive. We want to specify
fa2 contract address during print/origination, but the argparser isn't
pluggable.

Solution: Write a custom argparser and use that. Make use of the fact
that there's only one contract here to simplify. Also, add a convenience
command to originate the contract, and a convenience command to
print the git revision.